hook: Locate all source directories at startup
For now we'll stick with the default source directory in /etc, however
it is very likely we're going to extend this in future to support more
stateless directories and limit the number of files in /etc.

Our eventual goal is to make upstream apparmor stateless by default, and
allow applications to define *runtime* /etc/ snippets. So, we design our
source searcher with this in mind.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>