Implement our AoT compilation step
We walk each item in the source directories and call out to the helper
to compile all definitions one by one. In theory we can call the parser
with an entire directory but this has shown to be problematic when we
have one or more malformed profiles. This way we allow as much as possible
to compile and still retain flow and exit status.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>